refactor: HintItemService를 interface로 추상화
HintItem(Ball, Strike, Nothing)의 Service의 필수 메소드가 매우 유사함.
interface로 추상화를 통해, 추후 HintItem을 추가할 때 보다 편리하게 개발 가능.
패키지명 Hint->hint 오타 수정.
