+++ src/main/java/baseball/controller/BallController.java
@@ -1,8 +1,8 @@
 package baseball.controller;
 
 import baseball.domain.Baseball;
-import baseball.domain.Hint.Ball;
-import baseball.service.BallService;
+import baseball.domain.hint.Ball;
+import baseball.service.hint.BallService;
 
 public class BallController {
     private static final BallService ballService = BallService.getInstance();
@@ -19,4 +19,15 @@ public class BallController {
         return ball.isActive();
     }
 
+    public static String getCountIfOverValue(Ball ball, int value) {
+        String count = getEmptyIfIntNotOverValue(ball.getCount(), value);
+        return count;
+    }
+
+    public static String getEmptyIfIntNotOverValue(int i, int value){
+        if (i <= value) {
+            return "";
+        }
+        return String.valueOf(i);
+    }
 }
