+++ src/main/java/baseball/domain/Restart.java
@@ -15,6 +15,5 @@ public enum Restart {
         return value;
     }
 
-
 }
