+++ src/main/java/baseball/domain/hint/Ball.java
@@ -1,4 +1,4 @@
-package baseball.domain.Hint;
+package baseball.domain.hint;
 
 public class Ball extends HintItem {
