+++ src/main/java/baseball/domain/hint/Nothing.java
@@ -1,4 +1,4 @@
-package baseball.domain.Hint;
+package baseball.domain.hint;
 
 public class Nothing extends HintItem {
