+++ src/main/java/baseball/domain/hint/Strike.java
@@ -1,4 +1,4 @@
-package baseball.domain.Hint;
+package baseball.domain.hint;
 
 public class Strike extends HintItem{
