+++ src/test/java/baseball/service/BallServiceTest.java
@@ -1,6 +1,7 @@
 package baseball.service;
 
 import baseball.domain.Baseball;
+import baseball.service.hint.BallService;
 import baseball.util.IntegerListUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
